Almost fixed this piece of sh!
Added cmake changes so that whether library build and linked as DLL or not is handled by single command as it should have been from the start.
OpenGL calls (and inherently ImGui's calls to OpenGL and GLFW) would likely cause an access violation if the definition is in the header as an inline function or template. Can be fixed with initialization of glfw and glew on the clients side but there should be a better way, right? TODO: find the better way.